+++ app/src/main/java/com/shevchenkovtwo/homework/basemodels/Actor.kt
@@ -1,0 +1,11 @@
+package com.shevchenkovtwo.homework.basemodels
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class Actor(
+    val id: Int,
+    val name: String,
+    val picture: String?
+) : Parcelable
