+++ app/src/main/java/com/shevchenkovtwo/homework/basemodels/Configuration.kt
@@ -1,0 +1,3 @@
+package com.shevchenkovtwo.homework.basemodels
+
+data class Configuration(val imageUrl: String)
