+++ app/src/main/java/com/shevchenkovtwo/homework/basemodels/Genre.kt
@@ -1,13 +1,11 @@
-package com.shevchenkovtwo.homework.data.models
+package com.shevchenkovtwo.homework.basemodels
 
 import android.os.Parcelable
-import com.squareup.moshi.JsonClass
 import kotlinx.parcelize.Parcelize
 
 
-@JsonClass(generateAdapter = true)
 @Parcelize
 data class Genre(
     val id: Int,
     val name: String
-) : Parcelable
+):Parcelable
