+++ app/src/main/java/com/shevchenkovtwo/homework/basemodels/Movie.kt
@@ -1,0 +1,20 @@
+package com.shevchenkovtwo.homework.basemodels
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+
+@Parcelize
+data class Movie(
+    val id: Int,
+    val name: String,
+    val overview: String?,
+    val poster: String?,
+    val backdrop: String?,
+    val ratings: Float,
+    val numberOfRatings: Int,
+    val age: Int,
+    val runtime: Int?,
+    val genres: List<Genre>,
+    val actors: List<Actor>
+) : Parcelable
